lua: Use regular require instead of LoadModule
LoadModule just seems to do some convoluted require. It'll load the
module as text, pass it back down to lua for evaluation and then call it
as a function. All of this is done through loadfile and friends.

It's a table. Just fucking return it and require it.

Unless there's some magic here I'm too fucking stupid to understand.