chore: add more new tasks
Path handling is a mess, and I clarify that's due to me, not the
underlying code.

PoB seems to be a hybrid where it boots and if it doesn't find trees
it'll try to generate them. I think these steps should be separated. If
a tree is not found then we should just crash, not try to fix something
in runtime.