rt: Stub image loading and path fixes
How the lua sources are loaded is kinda confusing. From what I've
understood src directory is expanded to the root in the actual install.
Thus all the paths in the app are broken.

This is likely why they ship a built version of PoB to test locally as
you can't simply run from the tree.

These changes fix some of the loading. It's a shitty solution but it's a
step to the final direction.