rt: Add event loop through winit
This initializes the window for rendering and gives us an event loop to
use for other processing. Libuv for example could be plugged here.

Loop runs only once a second just to make developing easier. I rely on
the debug logs for now and they're not really readable at 180Hz.